Consolidated catalog and channel viewsets
